- 워닝 줄이도록 코드 고치기 - 코드 정리

+++ Source/TGON/Core/Drawing/Windows/WindowsDrawing.cpp
@@ -12,7 +12,7 @@ void DrawPoint(Window& window, const Point& p)
     HWND wndHandle = reinterpret_cast<HWND>(window.GetNativeWindow());
     HDC dcHandle = ::GetDC(wndHandle);
     {
-        SetPixel(dcHandle, p.x, p.y, RGB(0, 0, 0));
+        SetPixel(dcHandle, static_cast<int>(p.x), static_cast<int>(p.y), RGB(0, 0, 0));
     }
     ::ReleaseDC(wndHandle, dcHandle);
 }
@@ -22,8 +22,8 @@ void DrawLine(Window& window, const Point& p1, const Point& p2)
     HWND wndHandle = reinterpret_cast<HWND>(window.GetNativeWindow());
     HDC dcHandle = ::GetDC(wndHandle);
     {
-        ::MoveToEx(dcHandle, p1.x, p1.y, RGB(0, 0, 0));
-        ::LineTo(dcHandle, p2.x, p2.y);
+        ::MoveToEx(dcHandle, static_cast<int>(p1.x), static_cast<int>(p1.y), RGB(0, 0, 0));
+        ::LineTo(dcHandle, static_cast<int>(p2.x), static_cast<int>(p2.y));
     }
     ::ReleaseDC(wndHandle, dcHandle);
 }
@@ -32,9 +32,9 @@ void DrawPolygon(Window& window, const Point& p1, const Point& p2, const Point& 
 {
     const POINT ptTable[3] =
     {
-        {p1.x, p1.y},
-        {p2.x, p2.y},
-        {p3.x, p3.y},
+        {static_cast<LONG>(p1.x), static_cast<LONG>(p1.y)},
+        {static_cast<LONG>(p2.x), static_cast<LONG>(p2.y)},
+        {static_cast<LONG>(p3.x), static_cast<LONG>(p3.y)},
     };
 
     HWND wndHandle = reinterpret_cast<HWND>(window.GetNativeWindow());
